requirements.txt
appdirs==1.4.3
Django==2.0
django-phonenumber-field==2.0.0
djangorestframework==3.7.7
gunicorn==19.7.1
packaging==16.8
psycopg2==2.7.3.2
pyparsing==2.2.0
pytz==2017.3
six==1.11.0
